Separate diffusion and mass specific code from libceed.cpp.
