+++ fem/libceed_diffusion.cu
@@ -202,146 +202,3 @@ extern "C" __global__ void f_apply_diff(void *ctx, CeedInt Q,
     break;
   }
 }
-
-extern "C" __global__ void f_apply_diff_1d(void *ctx, CeedInt Q,
-                                        Fields_Cuda fields) {
-  // in[0], out[0] have shape [dim, nc=1, Q]
-  const CeedScalar *ug = (const CeedScalar *)fields.inputs[0];
-  const CeedScalar *qd = (const CeedScalar *)fields.inputs[1];
-  CeedScalar *vg = fields.outputs[0];
-  for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-       i < Q;
-       i += blockDim.x * gridDim.x) {
-    vg[i] = ug[i] * qd[i];
-  }
-}
-
-extern "C" __global__ void f_apply_diff_2d(void *ctx, CeedInt Q,
-                                        Fields_Cuda fields) {
-  // in[0], out[0] have shape [dim, nc=1, Q]
-  const CeedScalar *ug = (const CeedScalar *)fields.inputs[0];
-  const CeedScalar *qd = (const CeedScalar *)fields.inputs[1];
-  CeedScalar *vg = fields.outputs[0];
-  for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-       i < Q;
-       i += blockDim.x * gridDim.x) {
-    const CeedScalar ug0 = ug[i+Q*0];
-    const CeedScalar ug1 = ug[i+Q*1];
-    vg[i+Q*0] = qd[i+Q*0]*ug0 + qd[i+Q*1]*ug1;
-    vg[i+Q*1] = qd[i+Q*1]*ug0 + qd[i+Q*2]*ug1;
-  }
-}
-
-extern "C" __global__ void f_apply_diff_3d(void *ctx, CeedInt Q,
-                                        Fields_Cuda fields) {
-  // in[0], out[0] have shape [dim, nc=1, Q]
-  const CeedScalar *ug = (const CeedScalar *)fields.inputs[0];
-  const CeedScalar *qd = (const CeedScalar *)fields.inputs[1];
-  CeedScalar *vg = fields.outputs[0];
-  for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-       i < Q;
-       i += blockDim.x * gridDim.x) {
-    const CeedScalar ug0 = ug[i+Q*0];
-    const CeedScalar ug1 = ug[i+Q*1];
-    const CeedScalar ug2 = ug[i+Q*2];
-    vg[i+Q*0] = qd[i+Q*0]*ug0 + qd[i+Q*1]*ug1 + qd[i+Q*2]*ug2;
-    vg[i+Q*1] = qd[i+Q*1]*ug0 + qd[i+Q*3]*ug1 + qd[i+Q*4]*ug2;
-    vg[i+Q*2] = qd[i+Q*2]*ug0 + qd[i+Q*4]*ug1 + qd[i+Q*5]*ug2;
-  }
-}
-
-/// libCEED Q-function for building quadrature data for a mass operator
-extern "C" __global__ void f_build_mass_const(void *ctx, CeedInt Q,
-                        Fields_Cuda fields) {
-  // in[0] is Jacobians with shape [dim, nc=dim, Q]
-  // in[1] is quadrature weights, size (Q)
-  struct BuildContext *bc = (struct BuildContext*)ctx;
-  const CeedScalar coeff = bc->coeff;
-  const CeedScalar *J = (const CeedScalar *)fields.inputs[0];
-  const CeedScalar *qw = (const CeedScalar *)fields.inputs[1];
-  CeedScalar *qd = fields.outputs[0];
-  switch (bc->dim + 10*bc->space_dim) {
-  case 11:
-    for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-         i < Q;
-         i += blockDim.x * gridDim.x) {
-      qd[i] = coeff * J[i] * qw[i];
-    }
-    break;
-  case 22:
-    for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-         i < Q;
-         i += blockDim.x * gridDim.x) {
-      // 0 2
-      // 1 3
-      qd[i] = coeff * (J[i+Q*0]*J[i+Q*3] - J[i+Q*1]*J[i+Q*2]) * qw[i];
-    }
-    break;
-  case 33:
-    for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-         i < Q;
-         i += blockDim.x * gridDim.x) {
-      // 0 3 6
-      // 1 4 7
-      // 2 5 8
-      qd[i] = (J[i+Q*0]*(J[i+Q*4]*J[i+Q*8] - J[i+Q*5]*J[i+Q*7]) -
-               J[i+Q*1]*(J[i+Q*3]*J[i+Q*8] - J[i+Q*5]*J[i+Q*6]) +
-               J[i+Q*2]*(J[i+Q*3]*J[i+Q*7] - J[i+Q*4]*J[i+Q*6])) * coeff * qw[i];
-    }
-    break;
-  }
-}
-
-extern "C" __global__ void f_build_mass_grid(void *ctx, CeedInt Q,
-                        Fields_Cuda fields) {
-  // in[0] is Jacobians with shape [dim, nc=dim, Q]
-  // in[1] is quadrature weights, size (Q)
-  struct BuildContext *bc = (struct BuildContext*)ctx;
-  const CeedScalar *c = (const CeedScalar *)fields.inputs[0];
-  const CeedScalar *J = (const CeedScalar *)fields.inputs[1];
-  const CeedScalar *qw = (const CeedScalar *)fields.inputs[2];
-  CeedScalar *qd = fields.outputs[0];
-  switch (bc->dim + 10*bc->space_dim) {
-  case 11:
-    for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-         i < Q;
-         i += blockDim.x * gridDim.x) {
-      qd[i] = c[i] * J[i] * qw[i];
-    }
-    break;
-  case 22:
-    for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-         i < Q;
-         i += blockDim.x * gridDim.x) {
-      // 0 2
-      // 1 3
-      qd[i] = c[i] * (J[i+Q*0]*J[i+Q*3] - J[i+Q*1]*J[i+Q*2]) * qw[i];
-    }
-    break;
-  case 33:
-    for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-         i < Q;
-         i += blockDim.x * gridDim.x) {
-      // 0 3 6
-      // 1 4 7
-      // 2 5 8
-      qd[i] = (J[i+Q*0]*(J[i+Q*4]*J[i+Q*8] - J[i+Q*5]*J[i+Q*7]) -
-               J[i+Q*1]*(J[i+Q*3]*J[i+Q*8] - J[i+Q*5]*J[i+Q*6]) +
-               J[i+Q*2]*(J[i+Q*3]*J[i+Q*7] - J[i+Q*4]*J[i+Q*6])) * c[i] * qw[i];
-    }
-    break;
-  }
-}
-
-/// libCEED Q-function for applying a mass operator
-extern "C" __global__ void f_apply_mass(void *ctx, CeedInt Q,
-                        Fields_Cuda fields) {
-  const CeedScalar *u = (const CeedScalar *)fields.inputs[0];
-  const CeedScalar *w = (const CeedScalar *)fields.inputs[1];
-  CeedScalar *v = fields.outputs[0];
-  for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-       i < Q;
-       i += blockDim.x * gridDim.x) {
-    v[i] = w[i] * u[i];
-  }
-}
