+++ fem/libceed_diffusion.qf
@@ -147,9 +147,7 @@ inline __device__ void f_apply_diff(void *ctx, const CeedInt Q, const CeedScalar
   // in[0], out[0] have shape [dim, nc=1, Q]
   switch (bc->dim) {
   case 1:
-    for (int i = blockIdx.x * blockDim.x + threadIdx.x;
-         i < Q;
-         i += blockDim.x * gridDim.x) {
+    for (CeedInt i = 0; i < Q; ++i) {
       vg[i] = ug[i] * qd[i];
     }
     break;
@@ -174,100 +172,3 @@ inline __device__ void f_apply_diff(void *ctx, const CeedInt Q, const CeedScalar
   }
 }
 
-inline __device__ void f_apply_diff_1d(void *ctx, const CeedInt Q, const CeedScalar *ug, const CeedScalar *qd, CeedScalar *vg) {
-  // in[0], out[0] have shape [dim, nc=1, Q]
-  for (CeedInt i = 0; i < Q; ++i) {
-    vg[i] = ug[i] * qd[i];
-  }
-}
-
-inline __device__ void f_apply_diff_2d(void *ctx, const CeedInt Q, const CeedScalar *ug, const CeedScalar *qd, CeedScalar *vg) {
-  // in[0], out[0] have shape [dim, nc=1, Q]
-  for (CeedInt i = 0; i < Q; ++i) {
-    const CeedScalar ug0 = ug[i+Q*0];
-    const CeedScalar ug1 = ug[i+Q*1];
-    vg[i+Q*0] = qd[i+Q*0]*ug0 + qd[i+Q*1]*ug1;
-    vg[i+Q*1] = qd[i+Q*1]*ug0 + qd[i+Q*2]*ug1;
-  }
-}
-
-inline __device__ void f_apply_diff_3d(void *ctx, const CeedInt Q, const CeedScalar *ug, const CeedScalar *qd, CeedScalar *vg) {
-  // in[0], out[0] have shape [dim, nc=1, Q]
-  for (CeedInt i = 0; i < Q; ++i) {
-    const CeedScalar ug0 = ug[i+Q*0];
-    const CeedScalar ug1 = ug[i+Q*1];
-    const CeedScalar ug2 = ug[i+Q*2];
-    vg[i+Q*0] = qd[i+Q*0]*ug0 + qd[i+Q*1]*ug1 + qd[i+Q*2]*ug2;
-    vg[i+Q*1] = qd[i+Q*1]*ug0 + qd[i+Q*3]*ug1 + qd[i+Q*4]*ug2;
-    vg[i+Q*2] = qd[i+Q*2]*ug0 + qd[i+Q*4]*ug1 + qd[i+Q*5]*ug2;
-  }
-}
-
-/// libCEED Q-function for building quadrature data for a mass operator
-inline __device__ void f_build_mass_const(void *ctx, const CeedInt Q, const CeedScalar *J, const CeedScalar *qw, CeedScalar *qd) {
-  // in[0] is Jacobians with shape [dim, nc=dim, Q]
-  // in[1] is quadrature weights, size (Q)
-  struct BuildContext *bc = (struct BuildContext*)ctx;
-  const CeedScalar coeff = bc->coeff;
-  switch (bc->dim + 10*bc->space_dim) {
-  case 11:
-    for (CeedInt i = 0; i < Q; ++i) {
-      qd[i] = coeff * J[i] * qw[i];
-    }
-    break;
-  case 22:
-    for (CeedInt i = 0; i < Q; ++i) {
-      // 0 2
-      // 1 3
-      qd[i] = coeff * (J[i+Q*0]*J[i+Q*3] - J[i+Q*1]*J[i+Q*2]) * qw[i];
-    }
-    break;
-  case 33:
-    for (CeedInt i = 0; i < Q; ++i) {
-      // 0 3 6
-      // 1 4 7
-      // 2 5 8
-      qd[i] = (J[i+Q*0]*(J[i+Q*4]*J[i+Q*8] - J[i+Q*5]*J[i+Q*7]) -
-               J[i+Q*1]*(J[i+Q*3]*J[i+Q*8] - J[i+Q*5]*J[i+Q*6]) +
-               J[i+Q*2]*(J[i+Q*3]*J[i+Q*7] - J[i+Q*4]*J[i+Q*6])) * coeff * qw[i];
-    }
-    break;
-  }
-}
-
-inline __device__ void f_build_mass_grid(void *ctx, const CeedInt Q, const CeedScalar *c, const CeedScalar *J, const CeedScalar *qw, CeedScalar *qd) {
-  // in[0] is Jacobians with shape [dim, nc=dim, Q]
-  // in[1] is quadrature weights, size (Q)
-  struct BuildContext *bc = (struct BuildContext*)ctx;
-  switch (bc->dim + 10*bc->space_dim) {
-  case 11:
-    for (CeedInt i = 0; i < Q; ++i) {
-      qd[i] = c[i] * J[i] * qw[i];
-    }
-    break;
-  case 22:
-    for (CeedInt i = 0; i < Q; ++i) {
-      // 0 2
-      // 1 3
-      qd[i] = c[i] * (J[i+Q*0]*J[i+Q*3] - J[i+Q*1]*J[i+Q*2]) * qw[i];
-    }
-    break;
-  case 33:
-    for (CeedInt i = 0; i < Q; ++i) {
-      // 0 3 6
-      // 1 4 7
-      // 2 5 8
-      qd[i] = (J[i+Q*0]*(J[i+Q*4]*J[i+Q*8] - J[i+Q*5]*J[i+Q*7]) -
-               J[i+Q*1]*(J[i+Q*3]*J[i+Q*8] - J[i+Q*5]*J[i+Q*6]) +
-               J[i+Q*2]*(J[i+Q*3]*J[i+Q*7] - J[i+Q*4]*J[i+Q*6])) * c[i] * qw[i];
-    }
-    break;
-  }
-}
-
-/// libCEED Q-function for applying a mass operator
-inline __device__ void f_apply_mass(void *ctx, const CeedInt Q, const CeedScalar *u, const CeedScalar *w, CeedScalar *v) {
-  for (CeedInt i = 0; i < Q; ++i) {
-    v[i] = w[i] * u[i];
-  }
-}
